+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file
 
 ## [0.2.8] 2023-10-25
 
+### Added
+
+
+- Add method is board full and test
+
 ### Modify