+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file
 
+## [0.2.8] 2023-10-26
+
+### Changed
+
+
+- implement has letter
+
 ## [0.2.8] 2023-10-25
 
 ### Added